// src/app/shared/models/sorting.ts
export enum SortKeyWord {
  byId = 'user_id',
  byFirstName = 'first_name',
  byLastName = 'last_name',
  byBirthday = 'birthday',
  byPatName = 'pat_name',
  byPhone = 'phone',
  byDateLast = 'date_last',
  bySumm = 'summ',
  byBonus = 'bonus',
  byBarcode = 'barcode',
}

export enum BySort {
  ascending = '▼',
  descending = '▲',
}